Moving to conf system - BarChart
