+++ src/charts/i-bar-chart-conf.ts
@@ -1,0 +1,6 @@
+import {IStackMixinConf} from '../base/i-stack-mixin-conf';
+
+export interface IBarChartConf extends IStackMixinConf {
+    readonly alwaysUseRounding?: boolean;
+    readonly centerBar?: boolean;
+}
